fix: FixedTokenProvider 클래스 public 로 분리
